+++ .isort.cfg
@@ -1,7 +1,7 @@
 [settings]
 src_paths = .
 multi_line_output = 3
-known_third_party = decorator, jinja2, stripe, tornado, zulip
+known_third_party = decorator, jinja2, markdown, stripe, tornado, zulip
 include_trailing_comma = True
 use_parentheses = True
 line_length = 100
