+++ docs/documentation/user.md
@@ -248,9 +248,9 @@ languages in API docs, etc.  To create a tab switcher, write:
 
 The tab identifiers (e.g. `desktop-web` above) and their mappings to
 the tabs' labels are declared in
-[zerver/lib/bugdown/tabbed_sections.py][tabbed-sections-code].
+[zerver/lib/markdown/tabbed_sections.py][tabbed-sections-code].
 
-[tabbed-sections-code]: https://github.com/zulip/zulip/blob/master/zerver/lib/bugdown/tabbed_sections.py#L37
+[tabbed-sections-code]: https://github.com/zulip/zulip/blob/master/zerver/lib/markdown/tabbed_sections.py#L37
 
 This widget can also be used just to create a nice box around a set of
 instructions
