+++ mypy.ini
@@ -110,7 +110,7 @@ strict_optional = False
 strict_optional = False
 [mypy-zerver.lib.outgoing_webhook]
 strict_optional = False
-[mypy-zerver.lib.bugdown]  # for __init__.py
+[mypy-zerver.lib.markdown]  # for __init__.py
 strict_optional = False
 [mypy-zerver.lib.push_notifications]
 strict_optional = False
