+++ tools/test-backend
@@ -62,7 +62,7 @@ not_yet_fully_covered = {path for target in [
     'analytics/views.py',
     # Major lib files should have 100% coverage
     'zerver/lib/addressee.py',
-    'zerver/lib/bugdown/__init__.py',
+    'zerver/lib/markdown/__init__.py',
     'zerver/lib/cache.py',
     'zerver/lib/cache_helpers.py',
     'zerver/lib/i18n.py',
@@ -71,10 +71,10 @@ not_yet_fully_covered = {path for target in [
     'zerver/lib/url_preview/preview.py',
     'zerver/worker/queue_processors.py',
     # Bugdown sub-libs should have full coverage too; a lot are really close
-    'zerver/lib/bugdown/api_arguments_table_generator.py',
-    'zerver/lib/bugdown/fenced_code.py',
-    'zerver/lib/bugdown/help_relative_links.py',
-    'zerver/lib/bugdown/nested_code_blocks.py',
+    'zerver/lib/markdown/api_arguments_table_generator.py',
+    'zerver/lib/markdown/fenced_code.py',
+    'zerver/lib/markdown/help_relative_links.py',
+    'zerver/lib/markdown/nested_code_blocks.py',
     # Other lib files that ideally would coverage, but aren't sorted
     'zerver/__init__.py',
     'zerver/filters.py',
