+++ zerver/lib/import_realm.py
@@ -15,7 +15,7 @@ from psycopg2.extras import execute_values
 from psycopg2.sql import SQL, Identifier
 
 from analytics.models import RealmCount, StreamCount, UserCount
-from zerver.lib import bugdown
+from zerver.lib import markdown as bugdown
 from zerver.lib.actions import (
     UserMessageLite,
     bulk_insert_ums,
@@ -23,9 +23,9 @@ from zerver.lib.actions import (
     do_change_plan_type,
 )
 from zerver.lib.avatar_hash import user_avatar_path_from_ids
-from zerver.lib.bugdown import version as bugdown_version
 from zerver.lib.bulk_create import bulk_create_users, bulk_set_users_or_streams_recipient_fields
 from zerver.lib.export import DATE_FIELDS, Field, Path, Record, TableData, TableName
+from zerver.lib.markdown import version as bugdown_version
 from zerver.lib.parallel import run_parallel
 from zerver.lib.server_initialization import create_internal_realm, server_initialized
 from zerver.lib.streams import render_stream_description
