+++ zerver/lib/markdown/help_emoticon_translations_table.py
