+++ zerver/lib/markdown/tabbed_sections.py
