+++ zerver/views/report.py
@@ -9,7 +9,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 
 from zerver.decorator import human_users_only
-from zerver.lib.bugdown import privacy_clean_markdown
+from zerver.lib.markdown import privacy_clean_markdown
 from zerver.lib.queue import queue_json_publish
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
